🔍 refactor.input element 정리
